+++ spec/support/amazing_activist.rb
@@ -8,6 +8,38 @@ module Unconventional
   class ClassNaming < AmazingActivist::Base; end
 end
 
+class BrokenContractActivity < AmazingActivist::Base
+  class WithHandler < self
+    on_broken_outcome { |value| success(value) }
+  end
+
+  class WithHandlerOverride < WithHandler
+    on_broken_outcome do |value|
+      failure(:broken_contract, context: { value: value })
+    end
+  end
+
+  def call = 42
+end
+
+class UnhandledExceptionActivity < AmazingActivist::Base
+  class WithHandler < self
+    rescue_from("StandardError") do |exception|
+      failure(:unhandled_exception, exception: exception)
+    end
+  end
+
+  class WithHandlerOverride < WithHandler
+    rescue_from("RuntimeError") do
+      success(42)
+    end
+  end
+
+  def call
+    raise params.fetch(:error_class, RuntimeError), "Boom!"
+  end
+end
+
 module Pretty
   class DamnGoodActivity < AmazingActivist::Base
     def call
@@ -15,6 +47,24 @@ module Pretty
         success("YEAH! Let's go to the punk rock show!")
       else
         failure(:bad_choice)
+      end
+    end
+  end
+
+  class IrresistibleActivity < AmazingActivist::Base
+    def call
+      propose.unwrap!
+      success
+    end
+
+    private
+
+    def propose
+      case params[:proposal]
+      when :watch_tv_and_have_a_couple_of_brews
+        failure(:you_can_do_better)
+      else
+        raise "Oopsie!"
       end
     end
   end